Correctly position logo and credits
After much toying with flexboxes, the image credit, page credit,
and the logo banner are where I want them. In the end, I made a
blank <div> in the image-flexbox to make sure that there was an
element it could use to space-between the other items to just
appear in the middle and bottom of the page.