Check for password match
Using an event listener for input in the main.js, the webpage will
now tell the user in real time (by sending either blank or error
text to an empty div) weather or not the passwords match.

The password RegEx statement also successfully works and requires
the user to submit a password containing at least one number,
capital and lowercase letter.